requirements.txt
asgiref==3.8.1
Django==3.2
django-ckeditor==6.7.2
django-filter==21.1
django-js-asset==2.2.0
djangorestframework==3.11.0
Pillow==9.5.0
psycopg2-binary==2.8.6
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
